+++ public/assets/tarkov/images/grid_square.png
