+++ src/pages/tarkov/components/Listing/ListingItem.tsx
