+++ src/pages/tarkov/components/Listing/index.tsx
